move logic to activities
