+++ apps/github-agent/package.json
@@ -36,6 +36,7 @@
     "@temporalio/worker": "1.11.7",
     "@temporalio/workflow": "1.11.7",
     "@vertesia/client": "^0.50.1",
+    "@vertesia/common": "^0.55.0",
     "octokit": "4.1.2"
   }
 }
